requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
schedule==1.2.0
argon2_cffi==25.1.0
plotly
kafka-python==2.0.2